+++ src/core/core-config.c
@@ -3629,9 +3629,10 @@ config_weechat_init_options (void)
             /* TRANSLATORS: string "${color:xxx}" must NOT be translated. */
             N_("message displayed when the day has changed, with one date "
                "displayed (for example at beginning of buffer) (see man "
-               "strftime for date/time specifiers) (note: content is "
-               "evaluated, so you can use colors with format \"${color:xxx}\", "
-               "see /help eval)"),
+               "strftime for date/time specifiers, extra specifiers are "
+               "supported, see function util_strftimeval in Plugin API "
+               "reference) (note: content is evaluated, so you can use colors "
+               "with format \"${color:xxx}\", see /help eval)"),
             NULL, 0, 0, "-- %a, %d %b %Y --", NULL, 0,
             NULL, NULL, NULL,
             &config_change_buffers, NULL, NULL,